+++ benchmarks/versus-external-pools/utils.mjs
@@ -1,9 +1,0 @@
-export const executeAsyncFn = async fn => {
-  try {
-    await fn()
-  } catch (e) {
-    console.error(e)
-    // eslint-disable-next-line n/no-process-exit
-    process.exit(1)
-  }
-}