+++ benchmarks/versus-external-pools/workers/tinypool/function-to-bench-worker.mjs
@@ -1,2 +1,0 @@
-import functionToBench from '../../functions/function-to-bench.js'
-export default functionToBench