+++ prepare.js
@@ -1,4 +1,6 @@
-const isCIEnvironment = process.env.CI != null
+const { env } = require('node:process')
+
+const isCIEnvironment = env.CI != null
 if (isCIEnvironment === false) {
   require('husky').install()
 }