+++ tests/pools/thread/fixed.test.js
@@ -87,17 +87,21 @@ describe('Fixed thread pool test suite', () => {
         errorHandler: e => console.error(e)
       }
     )
+    expect(pool.emitter.eventNames()).toStrictEqual([])
     let poolReady = 0
     pool.emitter.on(PoolEvents.ready, () => ++poolReady)
     await waitPoolEvents(pool, PoolEvents.ready, 1)
+    expect(pool.emitter.eventNames()).toStrictEqual([PoolEvents.ready])
     expect(poolReady).toBe(1)
     await pool.destroy()
   })
 
   it("Verify that 'busy' event is emitted", async () => {
     const promises = new Set()
+    expect(pool.emitter.eventNames()).toStrictEqual([])
     let poolBusy = 0
     pool.emitter.on(PoolEvents.busy, () => ++poolBusy)
+    expect(pool.emitter.eventNames()).toStrictEqual([PoolEvents.busy])
     for (let i = 0; i < numberOfThreads * 2; i++) {
       promises.add(pool.execute())
     }
@@ -205,10 +209,12 @@ describe('Fixed thread pool test suite', () => {
 
   it('Verify that error handling is working properly:sync', async () => {
     const data = { f: 10 }
+    expect(errorPool.emitter.eventNames()).toStrictEqual([])
     let taskError
     errorPool.emitter.on(PoolEvents.taskError, e => {
       taskError = e
     })
+    expect(errorPool.emitter.eventNames()).toStrictEqual([PoolEvents.taskError])
     let inError
     try {
       await errorPool.execute(data)
@@ -234,10 +240,14 @@ describe('Fixed thread pool test suite', () => {
 
   it('Verify that error handling is working properly:async', async () => {
     const data = { f: 10 }
+    expect(asyncErrorPool.emitter.eventNames()).toStrictEqual([])
     let taskError
     asyncErrorPool.emitter.on(PoolEvents.taskError, e => {
       taskError = e
     })
+    expect(asyncErrorPool.emitter.eventNames()).toStrictEqual([
+      PoolEvents.taskError
+    ])
     let inError
     try {
       await asyncErrorPool.execute(data)
@@ -272,8 +282,13 @@ describe('Fixed thread pool test suite', () => {
 
   it('Shutdown test', async () => {
     const exitPromise = waitWorkerEvents(pool, 'exit', numberOfThreads)
+    expect(pool.emitter.eventNames()).toStrictEqual([PoolEvents.busy])
     let poolDestroy = 0
     pool.emitter.on(PoolEvents.destroy, () => ++poolDestroy)
+    expect(pool.emitter.eventNames()).toStrictEqual([
+      PoolEvents.busy,
+      PoolEvents.destroy
+    ])
     await pool.destroy()
     const numberOfExitEvents = await exitPromise
     expect(pool.started).toBe(false)