+++ CHANGELOG.md
@@ -7,6 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.7.4] - 2023-09-25
+
+### Fixed
+
+- Fix source maps (bundler issue).
+
+## [2.7.3] - 2023-09-24
+
+### Changed
+
+- Convert pool event emitter to event emitter async resource.
+
+## [2.7.2] - 2023-09-23
+
+### Changed
+
+- Add source maps to npm package to ease debugging.
+
+### Added
+
+- Continuous benchmarking versus other worker pools: [https://poolifier.github.io/benchmark](https://poolifier.github.io/benchmark).
+
 ## [2.7.1] - 2023-09-20
 
 ### Fixed
@@ -36,10 +58,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `startWorkers` to pool options to whether start the minimum number of workers at pool creation or not.
+- Add `startWorkers` to pool options to whether start the minimum number of workers at pool initialization or not.
 - Add `start()` method to pool API to start the minimum number of workers.
 - Add `taskStealing` and `tasksStealingOnPressure` to tasks queue options to whether enable task stealing or not and whether enable tasks stealing on back pressure or not.
-- Continuous internal benchmarking: https://poolifier.github.io/benchmark-results/dev/bench.
+- Continuous internal benchmarking: [https://poolifier.github.io/benchmark-results/dev/bench](https://poolifier.github.io/benchmark-results/dev/bench).
 
 ## [2.6.44] - 2023-09-08